*Huge refactor on build script
Mainly removed composer dependency from all targets. In default phing is installed via composer so we not need to do this installation again.

But composer task gets a huge rewrite that that allows to self-download composer if not present in the PATH. This should be useful if phing is installed globally and we only want to set-up project dependencies.